First implementation of atomic-instructions (for GCC & x86-64).
Includes CUnit-based unit-tests of the atomic instruction including running
with multiple threads.
These should be massaged to allow multiple tests run "per CI-suite"...

Quite some work put in to get the interfaces of the inlined functions
right, so that the compiler knows via __attribute__ whether parameters
are not null, are not used, etc.
TBD: including restrict of stack and bag-pointers (to not alias any other
pointer passed) -- and get it right with C++.